Implement logic to handle access

+++ library-system-be/src/main/java/com/myhomelibrary/library_system/controllers/LibraryBookCommentController.java
@@ -4,9 +4,11 @@ import com.myhomelibrary.library_system.domains.api.Response;
 import com.myhomelibrary.library_system.domains.comment.Comment;
 import com.myhomelibrary.library_system.domains.comment.CommentRequest;
 import com.myhomelibrary.library_system.domains.comment.CommentUpdateRequest;
+import com.myhomelibrary.library_system.repositories.CommentRepository;
+import com.myhomelibrary.library_system.repositories.LibraryRepository;
 import com.myhomelibrary.library_system.security.SecurityUtils;
 import com.myhomelibrary.library_system.services.CommentService;
-import jakarta.validation.Valid;
+import com.myhomelibrary.library_system.services.GenericAccessService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -21,6 +23,9 @@ import static com.myhomelibrary.library_system.controllers.LibraryController.LIB
 @AllArgsConstructor
 public class LibraryBookCommentController {
     private final CommentService commentService;
+    private final CommentRepository commentRepository;
+    private final LibraryRepository libraryRepository;
+    private final GenericAccessService genericAccessService;
 
     @GetMapping
     public Response<List<Comment>> getComments(@PathVariable UUID libraryId, @PathVariable UUID bookId) {
@@ -29,25 +34,28 @@ public class LibraryBookCommentController {
 
     @GetMapping("/{id}")
     public Response<Comment> getCommentById(@PathVariable UUID libraryId, @PathVariable UUID bookId, @PathVariable UUID id) {
+        genericAccessService.assertOwnerOrAdmin(libraryRepository::findLibraryById, libraryId);
         return Response.success(commentService.getCommentByIdInLibraryBook(libraryId, bookId, id));
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Response<Comment> createComment(@PathVariable UUID libraryId, @PathVariable UUID bookId, @RequestBody CommentRequest commentRequest) {
+        genericAccessService.assertOwnerOrAdmin(libraryRepository::findLibraryById, libraryId);
         return Response.success(commentService.createCommentInLibraryBook(libraryId, bookId, commentRequest, SecurityUtils.getAuthenticatedUserPk()));
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public UUID deleteComment(@PathVariable UUID libraryId, @PathVariable UUID bookId, @PathVariable UUID id) {
+        genericAccessService.assertOwnerOrAdmin(commentRepository::findCommentById, id);
         return commentService.deleteCommentByIdInLibraryBook(libraryId, bookId, id);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Response<Comment> updateComment(@PathVariable UUID libraryId, @PathVariable UUID bookId, @PathVariable UUID id, @RequestBody CommentUpdateRequest commentUpdateRequest) {
+        genericAccessService.assertOwnerOrAdmin(commentRepository::findCommentById, id);
         return Response.success(commentService.updateCommentInLibraryBook(libraryId, bookId, id, commentUpdateRequest));
     }
 }
-
