+++ library-system-be/src/main/java/com/myhomelibrary/library_system/controllers/LibraryController.java
@@ -3,13 +3,16 @@ package com.myhomelibrary.library_system.controllers;
 import com.myhomelibrary.library_system.domains.api.Response;
 import com.myhomelibrary.library_system.domains.library.Library;
 import com.myhomelibrary.library_system.domains.library.LibraryRequest;
-import com.myhomelibrary.library_system.services.LibraryService;
+import com.myhomelibrary.library_system.repositories.LibraryRepository;
 import com.myhomelibrary.library_system.security.SecurityUtils;
+import com.myhomelibrary.library_system.services.GenericAccessService;
+import com.myhomelibrary.library_system.services.LibraryService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
@@ -23,16 +26,24 @@ public class LibraryController {
     public static final String LIBRARY_BASE_URL = "/v1/library";
 
     private final LibraryService libraryService;
+    private final LibraryRepository libraryRepository;
+    private final GenericAccessService genericAccessService;
 
     @GetMapping
     public Response<Page<Library>> getLibraries(@PageableDefault(size = 20) Pageable pageable) {
-        // TODO: Only return user libraries? A separate endpoint for all libraries? Admin can access all
+        Long userPk = SecurityUtils.getAuthenticatedUserPk();
+        return Response.success(libraryService.getLibrariesByUserId(pageable, userPk));
+    }
+
+    @PreAuthorize("hasRole('ADMIN')")
+    @GetMapping("/all")
+    public Response<Page<Library>> getAllLibraries(@PageableDefault(size = 20) Pageable pageable) {
         return Response.success(libraryService.getAllLibraries(pageable));
     }
 
     @GetMapping("/{id}")
     public Response<Library> getLibraryById(@PathVariable UUID id) {
-        // TODO: Check if user has access to this library
+        genericAccessService.assertOwnerOrAdmin(libraryRepository::findLibraryById, id);
         return Response.success(libraryService.getLibraryById(id));
     }
 
@@ -45,14 +56,14 @@ public class LibraryController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Response<UUID> deleteLibrary(@PathVariable UUID id) {
-        // TODO: Check if user has access to delete this library
+        genericAccessService.assertOwnerOrAdmin(libraryRepository::findLibraryById, id);
         return Response.success(libraryService.deleteLibraryById(id));
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Response<Library> updateLibrary(@PathVariable UUID id, @RequestBody LibraryRequest libraryRequest) {
-        // TODO: Check if user has access to update this library. So maybe admin and the owner?
+        genericAccessService.assertOwnerOrAdmin(libraryRepository::findLibraryById, id);
         return Response.success(libraryService.updateLibrary(id, libraryRequest));
     }
 }
